refactor(blorktools): html to pages
