+++ packages/blorktools/src/asset_debugger/ui/pages/atlas-panel.html
