+++ packages/blorktools/src/asset_debugger/ui/pages/examples-modal.html
@@ -1,0 +1,18 @@
+<!-- Load Example Modal -->
+<div id="load-example-modal" class="modal-overlay">
+    <div class="modal-container">
+        <div class="modal-header">
+            <h3 class="modal-title">LOAD EXAMPLE</h3>
+            <button class="modal-close" id="close-example-modal">&times;</button>
+        </div>
+        <div class="modal-body" style="padding: 25px 20px;">
+            <div style="text-align: center; margin-bottom: 20px;">
+                <p style="margin-bottom: 20px; font-size: 16px;">SELECT AN EXAMPLE TO LOAD:</p>
+                <button id="rig-example-button">RIG EXAMPLE</button>
+            </div>
+            <div style="font-size: 12px; margin-top: 30px; color: #999; text-align: center;">
+                <p>EXAMPLES WILL LOAD WITHOUT REQUIRING ANY FILE UPLOADS</p>
+            </div>
+        </div>
+    </div>
+</div>
