+++ packages/blorktools/src/asset_debugger/ui/pages/mesh-panel.html
@@ -1,0 +1,5 @@
+<!-- Mesh Panel Component -->
+<div id="mesh-tab" class="tab-content active">
+    <h3>Mesh Visibility</h3>
+    <div id="mesh-groups"></div>
+</div>
