+++ packages/blorktools/src/asset_debugger/ui/pages/rig-panel.html
@@ -1,0 +1,6 @@
+<!-- Rig Panel Component -->
+<div id="rig-tab" class="tab-content">
+    <div id="rig-content" class="rig-info-container">
+        <p>No rig data found.</p>
+    </div>
+</div>
